refactor: move source resolution into api/services/source.py
ResolveError/Resolved + classify/resolve_local/resolve_source were defined
inline in the manifest router, but they're framework-agnostic domain logic (no
FastAPI), not Pydantic wire models — so they belong in the service layer beside
scan/cache/clone/media, not in models/ (the OpenAPI-serialization layer) and not
in the router. The router is now a thin handler that imports from the service;
the resolution exception lives next to the code that raises it (like
ScanCancelledError in scan.py, CloneError in clone.py). _resolve_local ->
resolve_local (now crosses a module boundary). No behavior change; backend 209
+ pyright 0, coverage retained via the existing route tests.
